+++ src/main/java/com/clova/anifriends/domain/applicant/dto/FindApplicantsResponse.java
@@ -2,7 +2,7 @@ package com.clova.anifriends.domain.applicant.dto;
 
 import com.clova.anifriends.domain.applicant.Applicant;
 import com.clova.anifriends.domain.recruitment.Recruitment;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerGender;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerGender;
 import java.time.LocalDate;
 import java.util.List;
