+++ src/main/java/com/clova/anifriends/domain/volunteer/Volunteer.java
@@ -6,12 +6,12 @@ import com.clova.anifriends.domain.applicant.Applicant;
 import com.clova.anifriends.domain.common.BaseTimeEntity;
 import com.clova.anifriends.domain.common.CustomPasswordEncoder;
 import com.clova.anifriends.domain.volunteer.exception.VolunteerBadRequestException;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerEmail;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerGender;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerName;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerPassword;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerPhoneNumber;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerTemperature;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerEmail;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerGender;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerName;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerPassword;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerPhoneNumber;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerTemperature;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
