+++ src/main/java/com/clova/anifriends/domain/volunteer/repository/VolunteerRepository.java
@@ -1,7 +1,7 @@
 package com.clova.anifriends.domain.volunteer.repository;
 
 import com.clova.anifriends.domain.volunteer.Volunteer;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerEmail;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerEmail;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
