+++ src/test/java/com/clova/anifriends/docs/enumtype/EnumDocumentationTest.java
@@ -12,7 +12,7 @@ import com.clova.anifriends.domain.animal.vo.AnimalType;
 import com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus;
 import com.clova.anifriends.domain.common.EnumType;
 import com.clova.anifriends.base.BaseControllerTest;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerGender;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerGender;
 import com.clova.anifriends.global.exception.ErrorCode;
 import java.util.Arrays;
 import java.util.Map;
