+++ src/test/java/com/clova/anifriends/domain/volunteer/service/VolunteerServiceTest.java
@@ -24,7 +24,7 @@ import com.clova.anifriends.domain.volunteer.repository.VolunteerRepository;
 import com.clova.anifriends.domain.volunteer.support.VolunteerDtoFixture;
 import com.clova.anifriends.domain.volunteer.support.VolunteerFixture;
 import com.clova.anifriends.domain.volunteer.support.VolunteerImageFixture;
-import com.clova.anifriends.domain.volunteer.wrapper.VolunteerGender;
+import com.clova.anifriends.domain.volunteer.vo.VolunteerGender;
 import com.clova.anifriends.global.image.S3Service;
 import java.time.LocalDate;
 import java.util.List;
